updated has_remaining_performs_key to use set
has_remaining_performs_key was using two commands to try and establish a lock.  In production I found this to be error prone.  I changed the code to use the latest form set which will set an expiration and only set the key if it doesn't exist.  This is the preferred way to establish a lock using redis per the documentation. http://redis.io/commands/set